@@ -1,15 +1,13 @@
-// utils/cloudinary.js
-const cloudinary = require('cloudinary').v2;
+import { v2 as cloudinary } from 'cloudinary';
+import dotenv from 'dotenv';
 
-// Function to connect to Cloudinary
-const cloudinaryConnect = () => {
-    cloudinary.config({
-        cloud_name: process.env.CLOUD_NAME,
-        api_key: process.env.API_KEY,
-        api_secret: process.env.API_SECRET,
-    });
-    console.log('Cloudinary connection established');
-};
+dotenv.config();
 
-// Export the connect function
-module.exports = { cloudinaryConnect };
+// Configure Cloudinary
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET
+});
+
+export default cloudinary;
