@@ -1,122 +1,218 @@
-const User = require('../models/User');
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const validator = require('validator');
+import User from '../models/User.js';
+import generateToken from '../utils/generateToken.js';
+import { validationResult } from 'express-validator';
+import { sendWelcomeEmail } from '../utils/emailService.js';
 
-
-const registerUser = async (req, res) => {
-    const { name, email, password, role } = req.body;
-
-    // Validate input fields
-    if (!name || !email || !password || !role) {
-        return res.status(400).json({ message: "All fields are required." });
+// @desc    Register user
+// @route   POST /api/v1/auth/register
+// @access  Public
+export const register = async (req, res) => {
+  try {
+    // Check for validation errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        success: false,
+        message: 'Validation failed',
+        errors: errors.array()
+      });
     }
 
-    if (!validator.isEmail(email)) {
-        return res.status(400).json({ message: "Please provide a valid email." });
+    const { name, email, password, role, phone, address } = req.body;
+
+    // Check if user exists
+    const userExists = await User.findOne({ email });
+    if (userExists) {
+      return res.status(400).json({
+        success: false,
+        message: 'User already exists with this email'
+      });
     }
 
-    if (password.length < 6) {
-        return res.status(400).json({ message: "Password must be at least 6 characters long." });
-    }
+    // Create user
+    const user = await User.create({
+      name,
+      email,
+      password,
+      role: role || 'customer',
+      phone,
+      address
+    });
 
-    try {
-        // Check if the email already exists
-        const userExists = await User.findOne({ email });
+    if (user) {
+      // Send welcome email (disabled for development)
+    // try {
+    //   await sendWelcomeEmail(user.email, user.name);
+    // } catch (emailError) {
+    //   console.error('Error sending welcome email:', emailError);
+    //   // Don't fail registration if email fails
+    // } 
 
-        if (userExists) {
-            return res.status(400).json({ message: "Email already in use." });
+      res.status(201).json({
+        success: true,
+        message: 'User registered successfully',
+        data: {
+          user: user.toJSON(),
+          token: generateToken(user._id)
         }
-
-        // Hash the password before saving
-        const salt = await bcrypt.genSalt(10);
-        const hashedPassword = await bcrypt.hash(password, salt);
-
-        // Create a new user
-        const newUser = new User({
-            name,
-            email,
-            password: hashedPassword,
-            role,
-        });
-
-        // Save the user to the database
-        await newUser.save();
-
-        // Generate a JWT token
-      
-
-        // Respond with a success message and the JWT token
-        res.status(201).json({
-            message: "User registered successfully!",
-        
-            user: {
-                id: newUser._id,
-                name: newUser.name,
-                email: newUser.email,
-                role: newUser.role,
-            }
-        });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: "Server error. Please try again later." });
+      });
+    } else {
+      res.status(400).json({
+        success: false,
+        message: 'Invalid user data'
+      });
     }
+  } catch (error) {
+    console.error('Register error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error during registration'
+    });
+  }
 };
 
 // @desc    Login user
-// @route   POST /api/auth/login
+// @route   POST /api/v1/auth/login
 // @access  Public
-const loginUser = async (req, res) => {
+export const login = async (req, res) => {
+  try {
+    // Check for validation errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        success: false,
+        message: 'Validation failed',
+        errors: errors.array()
+      });
+    }
+
     const { email, password } = req.body;
 
-    if (!email || !password) {
-        return res.status(400).json({ message: "Please provide email and password." });
+    // Check for user
+    const user = await User.findOne({ email }).select('+password');
+    if (!user) {
+      return res.status(401).json({
+        success: false,
+        message: 'Invalid credentials'
+      });
     }
 
-    try {
-        // Check if the user exists
-        const user = await User.findOne({ email });
-
-        if (!user) {
-            return res.status(400).json({ message: "Invalid credentials." });
-        }
-
-        // Compare the password with the hashed password
-        const isMatch = await bcrypt.compare(password, user.password);
-
-        if (!isMatch) {
-            return res.status(400).json({ message: "Invalid credentials." });
-        }
-
-        // Generate JWT token
-      // Example from login route:
-const token = jwt.sign(
-    { id: user._id, role: user.role },
-    process.env.JWT_SECRET,
-    { expiresIn: '15d' }
-);
-
-const options = {
-    expires: new Date(Date.now() + 3 * 24 * 60 * 60 * 1000),  // 3 days expiration
-    httpOnly: true,  
-    secure: process.env.NODE_ENV === 'production',  // true for production, false for local dev
-    sameSite: 'None'  // For cross-origin requests
-};
-
-// Sending the token in a cookie
-res.cookie('token', token, options).status(200).json({
-    success: true,
-    token,
-    user,
-    message: 'User Login Success',
-});
-
-        
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: "Server error. Please try again later." });
+    // Check password
+    const isMatch = await user.comparePassword(password);
+    if (!isMatch) {
+      return res.status(401).json({
+        success: false,
+        message: 'Invalid credentials'
+      });
     }
+
+    res.json({
+      success: true,
+      message: 'Login successful',
+      data: {
+        user: user.toJSON(),
+        token: generateToken(user._id)
+      }
+    });
+  } catch (error) {
+    console.error('Login error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error during login'
+    });
+  }
 };
 
+// @desc    Get current user
+// @route   GET /api/v1/auth/me
+// @access  Private
+export const getMe = async (req, res) => {
+  try {
+    const user = await User.findById(req.user.id);
+    
+    res.json({
+      success: true,
+      data: {
+        user: user.toJSON()
+      }
+    });
+  } catch (error) {
+    console.error('Get me error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
+};
 
-module.exports = { registerUser, loginUser };
+// @desc    Update user profile
+// @route   PUT /api/v1/auth/profile
+// @access  Private
+export const updateProfile = async (req, res) => {
+  try {
+    const { name, phone, address, avatar } = req.body;
+
+    const user = await User.findByIdAndUpdate(
+      req.user.id,
+      {
+        name: name || req.user.name,
+        phone: phone || req.user.phone,
+        address: address || req.user.address,
+        avatar: avatar || req.user.avatar
+      },
+      {
+        new: true,
+        runValidators: true
+      }
+    );
+
+    res.json({
+      success: true,
+      message: 'Profile updated successfully',
+      data: {
+        user: user.toJSON()
+      }
+    });
+  } catch (error) {
+    console.error('Update profile error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error during profile update'
+    });
+  }
+};
+
+// @desc    Change password
+// @route   PUT /api/v1/auth/change-password
+// @access  Private
+export const changePassword = async (req, res) => {
+  try {
+    const { currentPassword, newPassword } = req.body;
+
+    const user = await User.findById(req.user.id).select('+password');
+
+    // Check current password
+    const isMatch = await user.comparePassword(currentPassword);
+    if (!isMatch) {
+      return res.status(400).json({
+        success: false,
+        message: 'Current password is incorrect'
+      });
+    }
+
+    // Update password
+    user.password = newPassword;
+    await user.save();
+
+    res.json({
+      success: true,
+      message: 'Password changed successfully'
+    });
+  } catch (error) {
+    console.error('Change password error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error during password change'
+    });
+  }
+};
