@@ -1,174 +1,264 @@
-const Cart = require('../models/Cart.js');
-const Fabric = require('../models/Fabric.js');
-const Tailor = require('../models/Tailor.js');
+import Cart from '../models/Cart.js';
+import Fabric from '../models/Fabric.js';
+import { validationResult } from 'express-validator';
 
-// @desc    Add an item to the cart
-// @route   POST /api/cart
-// @access  Private (Only Customers can manage their carts)
-const addItemToCart = async (req, res) => {
-    const { customer, fabricId, quantity, tailorId, measurements, price } = req.body;
+// @desc    Get user cart
+// @route   GET /api/v1/cart
+// @access  Private
+export const getCart = async (req, res) => {
+  try {
+    let cart = await Cart.findOne({ user: req.user.id })
+      .populate('items.fabric', 'name image price stock shop category color')
+      .populate('items.fabric.shop', 'name city');
 
-    try {
-        // Check if the fabric exists
-        const fabric = await Fabric.findById(fabricId);
-        if (!fabric) {
-            return res.status(404).json({ message: 'Fabric not found' });
-        }
-
-        // Create or update the cart
-        let cart = await Cart.findOne({ customer });
-
-        if (!cart) {
-            // If no cart exists for the customer, create a new cart
-            cart = new Cart({
-                customer,
-                items: [{ fabric: fabricId, quantity }],
-                tailoringService: {
-                    tailor: tailorId,
-                    measurements,
-                    price
-                },
-                totalPrice: fabric.price * quantity + price
-            });
-        } else {
-            // If the cart exists, update it
-            const existingItem = cart.items.find(item => item.fabric.toString() === fabricId);
-            if (existingItem) {
-                existingItem.quantity += quantity; // Increase quantity
-            } else {
-                cart.items.push({ fabric: fabricId, quantity });
-            }
-
-            // Update tailoring service
-            cart.tailoringService.tailor = tailorId;
-            cart.tailoringService.measurements = measurements;
-            cart.tailoringService.price = price;
-
-            // Recalculate total price
-            cart.totalPrice = cart.items.reduce((total, item) => total + (item.quantity * fabric.price), 0) + price;
-        }
-
-        await cart.save();
-        res.status(200).json(cart);
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+    if (!cart) {
+      cart = await Cart.create({ user: req.user.id, items: [] });
     }
+
+    res.json({
+      success: true,
+      data: cart
+    });
+  } catch (error) {
+    console.error('Get cart error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
-// @desc    Get the cart of a customer
-// @route   GET /api/cart/:customerId
-// @access  Private (Only Customers can view their cart)
-const getCartByCustomer = async (req, res) => {
-    const { customerId } = req.params;
-
-    try {
-        const cart = await Cart.findOne({ customer: customerId }).populate('items.fabric').populate('tailoringService.tailor');
-
-        if (!cart) {
-            return res.status(404).json({ message: 'Cart not found' });
-        }
-
-        res.status(200).json(cart);
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+// @desc    Add item to cart
+// @route   POST /api/v1/cart/add
+// @access  Private
+export const addToCart = async (req, res) => {
+  try {
+    // Check for validation errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        success: false,
+        message: 'Validation failed',
+        errors: errors.array()
+      });
     }
+
+    const { fabricId, quantity = 1 } = req.body;
+
+    // Check if fabric exists and is available
+    const fabric = await Fabric.findById(fabricId);
+    if (!fabric) {
+      return res.status(404).json({
+        success: false,
+        message: 'Fabric not found'
+      });
+    }
+
+    if (!fabric.isActive) {
+      return res.status(400).json({
+        success: false,
+        message: 'Fabric is not available'
+      });
+    }
+
+    if (fabric.stock < quantity) {
+      return res.status(400).json({
+        success: false,
+        message: `Insufficient stock. Available: ${fabric.stock}`
+      });
+    }
+
+    // Find or create cart
+    let cart = await Cart.findOne({ user: req.user.id });
+    if (!cart) {
+      cart = new Cart({ user: req.user.id, items: [] });
+    }
+
+    // Check if item already exists in cart
+    const existingItemIndex = cart.items.findIndex(
+      item => item.fabric.toString() === fabricId
+    );
+
+    if (existingItemIndex > -1) {
+      // Update existing item quantity
+      const newQuantity = cart.items[existingItemIndex].quantity + quantity;
+      
+      if (newQuantity > fabric.stock) {
+        return res.status(400).json({
+          success: false,
+          message: `Cannot add more items. Total would exceed available stock (${fabric.stock})`
+        });
+      }
+
+      cart.items[existingItemIndex].quantity = newQuantity;
+      cart.items[existingItemIndex].subtotal = fabric.price * newQuantity;
+    } else {
+      // Add new item to cart
+      cart.items.push({
+        fabric: fabricId,
+        fabricId: fabricId,
+        quantity,
+        price: fabric.price,
+        subtotal: fabric.price * quantity
+      });
+    }
+
+    await cart.save();
+    await cart.populate('items.fabric', 'name image price stock shop category color');
+    await cart.populate('items.fabric.shop', 'name city');
+
+    res.json({
+      success: true,
+      message: 'Item added to cart successfully',
+      data: cart
+    });
+  } catch (error) {
+    console.error('Add to cart error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
-// @desc    Remove an item from the cart
-// @route   DELETE /api/cart/:cartId/item/:itemId
-// @access  Private (Only Customers can remove items from their cart)
-const removeItemFromCart = async (req, res) => {
-    const { cartId, itemId } = req.params;
-
-    try {
-        const cart = await Cart.findById(cartId);
-
-        if (!cart) {
-            return res.status(404).json({ message: 'Cart not found' });
-        }
-
-        const itemIndex = cart.items.findIndex(item => item._id.toString() === itemId);
-
-        if (itemIndex === -1) {
-            return res.status(404).json({ message: 'Item not found in cart' });
-        }
-
-        cart.items.splice(itemIndex, 1);
-        await cart.save();
-
-        res.status(200).json({ message: 'Item removed from cart', cart });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
-    }
-};
-
-// @desc    Update quantity of an item in the cart
-// @route   PUT /api/cart/:cartId/item/:itemId
-// @access  Private (Only Customers can update their cart items)
-const updateItemQuantity = async (req, res) => {
-    const { cartId, itemId } = req.params;
+// @desc    Update cart item quantity
+// @route   PUT /api/v1/cart/update/:fabricId
+// @access  Private
+export const updateCartItemQuantity = async (req, res) => {
+  try {
     const { quantity } = req.body;
+    const { fabricId } = req.params;
 
-    try {
-        const cart = await Cart.findById(cartId);
-
-        if (!cart) {
-            return res.status(404).json({ message: 'Cart not found' });
-        }
-
-        const item = cart.items.find(item => item._id.toString() === itemId);
-
-        if (!item) {
-            return res.status(404).json({ message: 'Item not found in cart' });
-        }
-
-        item.quantity = quantity;
-
-        // Recalculate total price
-        const totalPrice = cart.items.reduce((total, item) => total + (item.quantity * item.fabric.price), 0);
-        cart.totalPrice = totalPrice;
-
-        await cart.save();
-
-        res.status(200).json({ message: 'Item quantity updated', cart });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+    if (quantity < 1) {
+      return res.status(400).json({
+        success: false,
+        message: 'Quantity must be at least 1'
+      });
     }
-};
 
-// @desc    Clear the cart (remove all items)
-// @route   DELETE /api/cart/:cartId
-// @access  Private (Only Customers can clear their cart)
-const clearCart = async (req, res) => {
-    const { cartId } = req.params;
-
-    try {
-        const cart = await Cart.findById(cartId);
-
-        if (!cart) {
-            return res.status(404).json({ message: 'Cart not found' });
-        }
-
-        cart.items = [];
-        cart.totalPrice = 0;
-        await cart.save();
-
-        res.status(200).json({ message: 'Cart cleared' });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+    // Check fabric availability
+    const fabric = await Fabric.findById(fabricId);
+    if (!fabric) {
+      return res.status(404).json({
+        success: false,
+        message: 'Fabric not found'
+      });
     }
+
+    if (quantity > fabric.stock) {
+      return res.status(400).json({
+        success: false,
+        message: `Insufficient stock. Available: ${fabric.stock}`
+      });
+    }
+
+    const cart = await Cart.findOne({ user: req.user.id });
+    if (!cart) {
+      return res.status(404).json({
+        success: false,
+        message: 'Cart not found'
+      });
+    }
+
+    const itemIndex = cart.items.findIndex(
+      item => item.fabric.toString() === fabricId
+    );
+
+    if (itemIndex === -1) {
+      return res.status(404).json({
+        success: false,
+        message: 'Item not found in cart'
+      });
+    }
+
+    // Update item
+    cart.items[itemIndex].quantity = quantity;
+    cart.items[itemIndex].subtotal = fabric.price * quantity;
+
+    await cart.save();
+    await cart.populate('items.fabric', 'name image price stock shop category color');
+    await cart.populate('items.fabric.shop', 'name city');
+
+    res.json({
+      success: true,
+      message: 'Cart updated successfully',
+      data: cart
+    });
+  } catch (error) {
+    console.error('Update cart error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
-module.exports = {
-    addItemToCart,
-    getCartByCustomer,
-    removeItemFromCart,
-    updateItemQuantity,
-    clearCart
+// @desc    Remove item from cart
+// @route   DELETE /api/v1/cart/remove/:fabricId
+// @access  Private
+export const removeFromCart = async (req, res) => {
+  try {
+    const { fabricId } = req.params;
+
+    const cart = await Cart.findOne({ user: req.user.id });
+    if (!cart) {
+      return res.status(404).json({
+        success: false,
+        message: 'Cart not found'
+      });
+    }
+
+    // Remove item from cart
+    cart.items = cart.items.filter(
+      item => item.fabric.toString() !== fabricId
+    );
+
+    await cart.save();
+    await cart.populate('items.fabric', 'name image price stock shop category color');
+    await cart.populate('items.fabric.shop', 'name city');
+
+    res.json({
+      success: true,
+      message: 'Item removed from cart successfully',
+      data: cart
+    });
+  } catch (error) {
+    console.error('Remove from cart error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
+// @desc    Clear cart
+// @route   DELETE /api/v1/cart/clear
+// @access  Private
+export const clearCart = async (req, res) => {
+  try {
+    const cart = await Cart.findOne({ user: req.user.id });
+    if (!cart) {
+      return res.status(404).json({
+        success: false,
+        message: 'Cart not found'
+      });
+    }
+
+    cart.items = [];
+    cart.totalItems = 0;
+    cart.totalAmount = 0;
+    await cart.save();
+
+    res.json({
+      success: true,
+      message: 'Cart cleared successfully',
+      data: cart
+    });
+  } catch (error) {
+    console.error('Clear cart error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
+};
