@@ -1,120 +1,274 @@
-const Tailor = require('../models/Tailor');
-
-// @desc    Add a new tailor
-// @route   POST /api/tailors
-// @access  Private (Only Admin or Seller can add a tailor)
-const addTailor = async (req, res) => {
-    const { name, experience, skills, pricePerHour, description, city, contactNumber} = req.body;
-
-    try {
-        // Create a new tailor
-        const newTailor = new Tailor({
-            name,
-            experience,
-            skills,
-            pricePerHour,
-            description,
-            city,
-            contactNumber,
-        
-        });
-
-        // Save the tailor to the database
-        await newTailor.save();
-
-        res.status(201).json({ message: 'Tailor added successfully!', tailor: newTailor });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
-    }
-};
+import Tailor from '../models/Tailor.js';
+import { validationResult } from 'express-validator';
 
 // @desc    Get all tailors
-// @route   GET /api/tailors
+// @route   GET /api/v1/tailors
 // @access  Public
-const getAllTailors = async (req, res) => {
-    try {
-        const tailors = await Tailor.find();
-        res.status(200).json(tailors);
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+export const getAllTailors = async (req, res) => {
+  try {
+    const { 
+      city, 
+      specialization, 
+      minRating, 
+      search, 
+      sortBy = 'createdAt', 
+      page = 1, 
+      limit = 10 
+    } = req.query;
+
+    // Build query
+    let query = { isActive: true };
+    
+    if (city) {
+      query.city = { $regex: city, $options: 'i' };
     }
+    
+    if (specialization) {
+      query.specialization = { $in: [specialization] };
+    }
+    
+    if (minRating) {
+      query.rating = { $gte: parseFloat(minRating) };
+    }
+    
+    if (search) {
+      query.$text = { $search: search };
+    }
+
+    // Execute query with pagination
+    const tailors = await Tailor.find(query)
+      .populate('owner', 'name email')
+      .sort({ [sortBy]: -1 })
+      .limit(limit * 1)
+      .skip((page - 1) * limit);
+
+    const total = await Tailor.countDocuments(query);
+
+    res.json({
+      success: true,
+      count: tailors.length,
+      total,
+      pagination: {
+        page: parseInt(page),
+        limit: parseInt(limit),
+        pages: Math.ceil(total / limit)
+      },
+      data: tailors
+    });
+  } catch (error) {
+    console.error('Get all tailors error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
-// @desc    Get tailor by ID
-// @route   GET /api/tailors/:id
+// @desc    Get single tailor
+// @route   GET /api/v1/tailors/:id
 // @access  Public
-const getTailorById = async (req, res) => {
-    const { id } = req.params;
+export const getTailor = async (req, res) => {
+  try {
+    const tailor = await Tailor.findById(req.params.id)
+      .populate('owner', 'name email phone')
+      .populate('reviews.user', 'name avatar');
 
-    try {
-        const tailor = await Tailor.findById(id);
-        
-        if (!tailor) {
-            return res.status(404).json({ message: 'Tailor not found.' });
-        }
-
-        res.status(200).json(tailor);
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+    if (!tailor) {
+      return res.status(404).json({
+        success: false,
+        message: 'Tailor not found'
+      });
     }
+
+    res.json({
+      success: true,
+      data: tailor
+    });
+  } catch (error) {
+    console.error('Get tailor error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
-// @desc    Update tailor by ID
-// @route   PUT /api/tailors/:id
-// @access  Private (Only Admin or Seller can update a tailor)
-const updateTailor = async (req, res) => {
-    const { id } = req.params;
-    const { name, experience, skills, pricePerHour, description, city, contactNumber, portfolio } = req.body;
-
-    try {
-        const tailor = await Tailor.findById(id);
-
-        if (!tailor) {
-            return res.status(404).json({ message: 'Tailor not found.' });
-        }
-
-        // Update tailor details
-        tailor.name = name || tailor.name;
-        tailor.experience = experience || tailor.experience;
-        tailor.skills = skills || tailor.skills;
-        tailor.pricePerHour = pricePerHour || tailor.pricePerHour;
-        tailor.description = description || tailor.description;
-        tailor.city = city || tailor.city;
-        tailor.contactNumber = contactNumber || tailor.contactNumber;
-   
-
-        await tailor.save();
-
-        res.status(200).json({ message: 'Tailor updated successfully!', tailor });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+// @desc    Create new tailor profile
+// @route   POST /api/v1/tailors
+// @access  Private (Tailor role only)
+export const createTailor = async (req, res) => {
+  try {
+    // Check for validation errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        success: false,
+        message: 'Validation failed',
+        errors: errors.array()
+      });
     }
+
+    // Check if user already has a tailor profile
+    const existingTailor = await Tailor.findOne({ owner: req.user.id });
+    if (existingTailor) {
+      return res.status(400).json({
+        success: false,
+        message: 'You already have a tailor profile'
+      });
+    }
+
+    // Add user to req.body
+    req.body.owner = req.user.id;
+
+    const tailor = await Tailor.create(req.body);
+
+    res.status(201).json({
+      success: true,
+      message: 'Tailor profile created successfully',
+      data: tailor
+    });
+  } catch (error) {
+    console.error('Create tailor error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
-// @desc    Delete tailor by ID
-// @route   DELETE /api/tailors/:id
-// @access  Private (Only Admin or Seller can delete a tailor)
-const deleteTailor = async (req, res) => {
-    const { id } = req.params;
+// @desc    Update tailor profile
+// @route   PUT /api/v1/tailors/:id
+// @access  Private (Tailor owner only)
+export const updateTailor = async (req, res) => {
+  try {
+    let tailor = await Tailor.findById(req.params.id);
 
-    try {
-        const tailor = await Tailor.findById(id);
-
-        if (!tailor) {
-            return res.status(404).json({ message: 'Tailor not found.' });
-        }
-
-        await tailor.remove();
-
-        res.status(200).json({ message: 'Tailor deleted successfully!' });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: 'Server error. Please try again later.' });
+    if (!tailor) {
+      return res.status(404).json({
+        success: false,
+        message: 'Tailor not found'
+      });
     }
+
+    // Make sure user is tailor owner
+    if (tailor.owner.toString() !== req.user.id && req.user.role !== 'admin') {
+      return res.status(401).json({
+        success: false,
+        message: 'Not authorized to update this tailor profile'
+      });
+    }
+
+    tailor = await Tailor.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+      runValidators: true
+    });
+
+    res.json({
+      success: true,
+      message: 'Tailor profile updated successfully',
+      data: tailor
+    });
+  } catch (error) {
+    console.error('Update tailor error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
 };
 
-module.exports = { addTailor, getAllTailors, getTailorById, updateTailor, deleteTailor };
+// @desc    Delete tailor profile
+// @route   DELETE /api/v1/tailors/:id
+// @access  Private (Tailor owner only)
+export const deleteTailor = async (req, res) => {
+  try {
+    const tailor = await Tailor.findById(req.params.id);
+
+    if (!tailor) {
+      return res.status(404).json({
+        success: false,
+        message: 'Tailor not found'
+      });
+    }
+
+    // Make sure user is tailor owner
+    if (tailor.owner.toString() !== req.user.id && req.user.role !== 'admin') {
+      return res.status(401).json({
+        success: false,
+        message: 'Not authorized to delete this tailor profile'
+      });
+    }
+
+    // Soft delete - just mark as inactive
+    tailor.isActive = false;
+    await tailor.save();
+
+    res.json({
+      success: true,
+      message: 'Tailor profile deleted successfully'
+    });
+  } catch (error) {
+    console.error('Delete tailor error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
+};
+
+// @desc    Add tailor review
+// @route   POST /api/v1/tailors/:id/reviews
+// @access  Private
+export const addTailorReview = async (req, res) => {
+  try {
+    const { rating, comment, images } = req.body;
+
+    const tailor = await Tailor.findById(req.params.id);
+
+    if (!tailor) {
+      return res.status(404).json({
+        success: false,
+        message: 'Tailor not found'
+      });
+    }
+
+    // Check if user already reviewed this tailor
+    const alreadyReviewed = tailor.reviews.find(
+      review => review.user.toString() === req.user.id.toString()
+    );
+
+    if (alreadyReviewed) {
+      return res.status(400).json({
+        success: false,
+        message: 'You have already reviewed this tailor'
+      });
+    }
+
+    const review = {
+      user: req.user.id,
+      customerName: req.user.name,
+      rating: Number(rating),
+      comment,
+      images: images || []
+    };
+
+    tailor.reviews.push(review);
+    tailor.totalReviews = tailor.reviews.length;
+
+    // Update tailor rating
+    tailor.rating = tailor.reviews.reduce((acc, item) => item.rating + acc, 0) / tailor.reviews.length;
+
+    await tailor.save();
+
+    res.status(201).json({
+      success: true,
+      message: 'Review added successfully'
+    });
+  } catch (error) {
+    console.error('Add tailor review error:', error);
+    res.status(500).json({
+      success: false,
+      message: 'Server error'
+    });
+  }
+};
